[Move] CustomScrollView - Resource/Components 로 이동 #29
